Fix #1000 RTMClient v2 still requires aiohttp installed (even though it's unused)
